Create README.md
In this code, we have two programs, producer and consumer, which will connect with each other. The producer program will generate items and put them onto the table. The consumer program will pick up items from the table. 
 
The table can only hold 2 items at the same time. When the table is full, the producer will wait. When there are no items on the table, the consumer will wait. We use semaphores to synchronize the producer and consumer. Mutual exclusion should be considered. We use pthreads in the producer program and consumer program. Shared memory is used for the "table".